refactor a bit for shorter module
