+++ tests_mem/test_mem.py
@@ -5,7 +5,7 @@ import wgpu.backends.rs
 
 import pytest
 from testutils import can_use_glfw, can_use_wgpu_lib, can_use_pyside6
-
+from testutils import create_and_release, get_counts, ob_name_from_test_func
 
 if not can_use_wgpu_lib:
     pytest.skip(
@@ -17,61 +17,8 @@ if not can_use_wgpu_lib:
 DEVICE = wgpu.utils.get_default_device()
 
 
-# %% The logic
-
-
-def get_counts():
-    """Get a dict that maps object names to a 2-tuple represening
-    the counts in py and wgpu-native.
-    """
-    counts_py = wgpu.diagnostics.object_counts.get_dict()
-    counts_native = wgpu.diagnostics.rs_counts.get_dict()
-
-    all_keys = set(counts_py) | set(counts_native)
-
-    default = {"count": -1}
-
-    counts = {}
-    for key in sorted(all_keys):
-        counts[key] = (
-            counts_py.get(key, default)["count"],
-            counts_native.get(key, default)["count"],
-        )
-    counts.pop("total")
-
-    return counts
-
-
-def get_excess_counts(counts1, counts2):
-    """Compare two counts dicts, and return a new dict with the fields
-    that have increased counts.
-    """
-    more = {}
-    for name in counts1:
-        c1 = counts1[name][0]
-        c2 = counts2[name][0]
-        more_py = 0
-        if c2 > c1:
-            more_py = c2 - c1
-        c1 = counts1[name][1]
-        c2 = counts2[name][1]
-        more_native = 0
-        if c2 > c1:
-            more_native = c2 - c1
-        if more_py or more_native:
-            more[name] = more_py, more_native
-    return more
-
-
-def ob_name_from_test_func(func):
-    """Translate test_release_bind_group() to "BindGroup"."""
-    func_name = func.__name__
-    prefix = "test_release_"
-    assert func_name.startswith(prefix)
-    words = func_name[len(prefix) :].split("_")
-    if words[-1].isnumeric():
-        words.pop(-1)
-    return "".join(word.capitalize() for word in words)
+async def stub_event_loop():
+    pass
 
 
 def make_draw_func_for_canvas(canvas):
@@ -101,80 +48,6 @@ def make_draw_func_for_canvas(canvas):
         ctx.present()
 
     return draw
-
-
-def create_and_release(create_objects_func):
-    """Decorator."""
-
-    def core_test_func():
-        """The core function that does the testing."""
-
-        n = 32
-
-        generator = create_objects_func(n)
-        ob_name = ob_name_from_test_func(create_objects_func)
-
-        # ----- Collect options
-
-        options = {
-            "expected_counts_after_create": {ob_name: (32, 32)},
-            "expected_counts_after_release": {},
-        }
-
-        func_options = next(generator)
-        assert isinstance(func_options, dict), "First yield must be an options dict"
-        options.update(func_options)
-
-        # Measure baseline object counts
-        counts1 = get_counts()
-
-        # ----- Create
-
-        # Create objects
-        objects = list(generator)
-
-        # Test the count
-        assert len(objects) == n
-
-        # Test that all objects are of the same class.
-        # (this for-loop is a bit weird, but its to avoid leaking refs to objects)
-        cls = objects[0].__class__
-        assert all(isinstance(objects[i], cls) for i in range(len(objects)))
-
-        # Test that class matches function name (should prevent a group of copy-paste errors)
-        assert ob_name == cls.__name__[3:]
-
-        # Measure peak object counts
-        counts2 = get_counts()
-        more2 = get_excess_counts(counts1, counts2)
-        print("  more after create:", more2)
-
-        # Make sure the actual object has increased
-        assert more2  # not empty
-        assert more2 == options["expected_counts_after_create"]
-
-        # It's ok if other objects are created too ...
-
-        # ----- Release
-
-        # Delete objects
-        del objects
-        gc.collect()
-
-        # Measure after-release object counts
-        counts3 = get_counts()
-        more3 = get_excess_counts(counts1, counts3)
-        print("  more after release:", more3)
-
-        # Check!
-        assert more3 == options["expected_counts_after_release"]
-
-    core_test_func.__name__ = create_objects_func.__name__
-    return core_test_func
-
-
-async def stub_event_loop():
-    pass
 
 
 # %% Meta tests
